# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Postbag API
  version: 0.1.0
  description: A form backend that routes. Websites POST submissions to
    /s/{formId} (public, unauthenticated); everything under /v1 manages forms,
    streams, destinations, routes and deliveries and requires a bearer API key.
    Start at GET /v1/me, then see /llms.txt for the full agent onboarding guide.
servers:
  - url: https://postbag.dev
  - url: https://api.postbag.dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: pb_live_… API key
      description: "An API key minted via POST /v1/api-keys (or the dashboard), sent
        as `Authorization: Bearer pb_live_…`. Scopes are manage ⊇ read ⊇ submit
        — a manage key satisfies every read- or submit-scoped call too."
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            hint:
              type: string
            docs:
              type: string
            details:
              type: object
              additionalProperties: {}
          required:
            - code
            - message
      required:
        - error
...
    Form:
      type: object
      properties:
        id:
          type: string
          example: fm_8f3kq2
        project_id:
          type: string
          example: fm_8f3kq2
        slug:
          type: string
        name:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum:
            - active
            - paused
        schema_mode:
          type: string
          enum:
            - observe
            - enforce
            - managed
        current_schema_version:
          type:
            - integer
            - "null"
        settings:
          type: object
          additionalProperties: {}
        submit_url:
          type: string
        streams:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: fm_8f3kq2
              slug:
                type: string
              mapping_status:
                type: string
                enum:
                  - valid
                  - incomplete
            required:
              - id
              - slug
              - mapping_status
        counts:
          type: object
          properties:
            submissions:
              type: integer
            last_submission_at:
              type:
                - string
                - "null"
              example: 2026-08-21T09:00:00.000Z
          required:
            - submissions
            - last_submission_at
        created_at:
          type: string
          example: 2026-08-21T09:00:00.000Z
      required:
        - id
        - project_id
        - slug
        - name
        - tags
        - status
        - schema_mode
        - current_schema_version
        - settings
        - submit_url
        - streams
        - counts
        - created_at
    Destination:
      type: object
      properties:
        id:
          type: string
          example: fm_8f3kq2
        type:
          type: string
          enum:
            - email
            - telegram
            - webhook
            - slack
            - discord
        name:
          type: string
        config:
          type: object
          additionalProperties: {}
        health:
          type: string
          enum:
            - unknown
            - ok
            - failing
        verified:
          type: boolean
        created_at:
          type: string
          example: 2026-08-21T09:00:00.000Z
      required:
        - id
        - type
        - name
        - config
        - health
        - verified
        - created_at
    Route:
      type: object
      properties:
        id:
          type: string
          example: fm_8f3kq2
        form_id:
          type:
            - string
            - "null"
        stream_id:
          type:
            - string
            - "null"
        destination_id:
          type: string
          example: fm_8f3kq2
        enabled:
          type: boolean
        mode:
          type: object
          additionalProperties: {}
        window:
          type: object
          additionalProperties: {}
        quality:
          type: object
          additionalProperties: {}
        filter:
          type:
            - string
            - "null"
        transform:
          type:
            - string
            - "null"
        counts:
          type: object
          additionalProperties: {}
        created_at:
          type: string
          example: 2026-08-21T09:00:00.000Z
      required:
        - id
        - form_id
        - stream_id
        - destination_id
        - enabled
        - mode
    Embed:
      type: object
      properties:
        html:
          type: string
        fetch:
          type: string
        react:
          type: string
        astro:
          type: string
        nextjs_action:
          type: string
      required:
        - html
        - fetch
        - react
        - astro
        - nextjs_action
    Verify:
      type: object
      properties:
        curl:
          type: string
        then:
          type: string
      required:
        - curl
        - then
...
paths:
...
  /v1/billing/checkout:
    post:
      operationId: billing_checkout
      tags:
        - billing
      summary: Start a Polar checkout
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                plan:
                  type: string
                  enum:
                    - pro
                    - team
                interval:
                  type: string
                  enum:
                    - month
                    - year
              required:
                - plan
                - interval
      responses:
        "201":
          description: checkout created
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                required:
                  - url
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/billing/portal:
    post:
      operationId: billing_portal
      tags:
        - billing
      summary: Open the Polar customer portal
      responses:
        "201":
          description: portal session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                required:
                  - url
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/quickstart:
    post:
      operationId: quickstart
      tags:
        - discovery
      summary: One call to a working, routed form
      description: Creates (idempotently, by name within project) the project if
        missing, the form, a destination for whichever of
        notify_email/telegram/webhook was given, and a direct route, then
        returns the form, an embeddable snippet, a curl command to verify
        delivery, and next steps. Everything it does is also available as
        individual calls (POST /v1/forms, /v1/destinations, /v1/routes); this is
        a convenience, not a special path.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                project:
                  type: string
                  default: default
                origin:
                  type: string
                  format: uri
                notify_email:
                  type: string
                  format: email
                telegram:
                  type: object
                  properties:
                    bot_token:
                      type: string
                    chat_id:
                      type: string
                  required:
                    - bot_token
                    - chat_id
                webhook:
                  type: object
                  properties:
                    url:
                      type: string
                      format: uri
                    secret:
                      type: string
                  required:
                    - url
                tags:
                  type: array
                  items:
                    type: string
                  default: []
                redirect_url:
                  type: string
                  format: uri
              required:
                - name
      responses:
        "201":
          description: created (or returned)
          content:
            application/json:
              schema:
                type: object
                properties:
                  form:
                    $ref: "#/components/schemas/Form"
                  destinations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Destination"
                  routes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Route"
                  embed:
                    $ref: "#/components/schemas/Embed"
                  verify:
                    $ref: "#/components/schemas/Verify"
                  next:
                    type: array
                    items:
                      type: object
                      properties:
                        why:
                          type: string
                        method:
                          type: string
                        path:
                          type: string
                        body:
                          type: object
                          additionalProperties: {}
                required:
                  - form
                  - destinations
                  - routes
                  - embed
                  - verify
                  - next
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...